:hammer:新增头文件，分离代码
